+++ app/scripts/services/dimItemService.factory.js
@@ -22,70 +22,80 @@
           });
       }
 
-      function updateItemModel(item, source, target, equip) {
-        var matchingItem;
-
+      // TODO: Switch away from moveAmount property towards an explicit parameter like equip is
+      // Returns the new or updated item (it may create a new item!)
+      function updateItemModel(item, source, target, equip, amount) {
+        // If we've moved to a new place
         if (source.id !== target.id) {
-          var index = _.findIndex(source.items, function(i) {
-            return (item.index === i.index);
-          });
+          // We handle moving stackable and nonstackable items almost exactly the same!
+          var stackable = item.maxStackSize > 1;
+          // Items to be decremented
+          var sourceItems = stackable ?
+                _.sortBy(_.select(source.items, function(i) {
+                  return i.hash === item.hash &&
+                    i.id === item.id;
+                }), 'amount') : [item];
+          // Items to be incremented. There's really only ever at most one of these, but
+          // it's easier to deal with as a list.
+          var targetItems = stackable ?
+                _.sortBy(_.select(target.items, function(i) {
+                  return i.hash === item.hash &&
+                    i.id === item.id &&
+                    // Don't consider full stacks as targets
+                    i.amount !== i.maxStackSize;
+                }), 'amount') : [];
+          // moveAmount could be more than maxStackSize if there is more than one stack on a character!
+          var moveAmount = amount || item.amount;
+          var addAmount = moveAmount;
+          var removeAmount = moveAmount;
+          var removedSourceItem = false;
 
-          if (item.maxStackSize > 1 && item.amount < item.maxStackSize) { // Balance the stacks.
-            if (_.has(item, 'moveAmount') && (item.moveAmount > 0)) {
-              matchingItem = _.reduce(source.items, function(memo, i) {
-                if (item.hash === i.hash) {
-                  if (!(_.has(i, 'moveAmount')) || ((_.has(i, 'moveAmount') && i.moveAmount === 0))) {
-                    if (memo === null) {
-                      memo = i;
-                    } else if (memo.amount > i.amount) {
-                      memo = i;
-                    }
-                  }
-                }
-
-                return memo;
-              }, null);
-
-              if (!_.isNull(matchingItem)) {
-                if (item.moveAmount > item.amount) {
-                  matchingItem.amount = matchingItem.amount + (item.amount - item.moveAmount);
-                }
-              }
-
-              item.amount = item.moveAmount;
+          // Remove inventory from the source
+          while (removeAmount > 0) {
+            var sourceItem = sourceItems.shift();
+            if (!sourceItem) {
+              throw new Error("Looks like you requested to move more of this item than exists in the source!");
             }
 
-            item.moveAmount = 0;
-
-            matchingItem = _.filter(target.items, function(i) {
-              return ((i.amount < item.maxStackSize) && (i.hash === item.hash));
-            });
-
-            if (_.size(matchingItem) > 0) {
-              var mItem = matchingItem[0];
-              var combinedTotal = mItem.amount + item.amount;
-
-              item.moveAmount = item.amount;
-
-              if (combinedTotal <= item.maxStackSize) {
-                mItem.amount = combinedTotal;
-                item.amount = 0;
-              } else {
-                mItem.amount = mItem.maxStackSize;
-                item.amount = combinedTotal - item.maxStackSize;
+            var amountToRemove = Math.min(removeAmount, sourceItem.amount);
+            if (amountToRemove === sourceItem.amount) {
+              // Completely remove the source item
+              var sourceIndex = _.findIndex(source.items, function(i) {
+                return sourceItem.index === i.index;
+              });
+              if (sourceIndex >= 0) {
+                source.items.splice(sourceIndex, 1);
+                removedSourceItem = sourceItem.index === item.index;
               }
+            } else {
+              sourceItem.amount -= amountToRemove;
             }
+
+            removeAmount -= amountToRemove;
           }
 
-          item.owner = target.id;
+          // Add inventory to the target (destination)
+          var targetItem;
+          while (addAmount > 0) {
+            targetItem = targetItems.shift();
 
-          if (index >= 0) {
-            source.items.splice(index, 1);
-          }
+            if (!targetItem) {
+              targetItem = item;
+              if (!removedSourceItem) {
+                targetItem = angular.copy(item);
+                targetItem.index = dimStoreService.createItemIndex(targetItem);
+              }
+              removedSourceItem = false; // only move without cloning once
+              targetItem.amount = 0; // We'll increment amount below
+              target.items.push(targetItem);
+              targetItem.owner = target.id;
+            }
 
-          if (item.amount > 0) {
-            target.items.push(item);
+            var amountToAdd = Math.min(addAmount, targetItem.maxStackSize - targetItem.amount);
+            targetItem.amount += amountToAdd;
+            addAmount -= amountToAdd;
           }
+          item = targetItem; // The item we're operating on switches to the last target
         }
 
         if (equip) {
@@ -173,9 +183,11 @@
 
       function equipItem(item) {
         return dimBungieService.equip(item)
-          .then(dimStoreService.getStore.bind(null, item.owner))
+          .then(function() {
+            return dimStoreService.getStore(item.owner);
+          })
           .then(function(store) {
-            updateItemModel(item, store, store, true);
+            return updateItemModel(item, store, store, true);
           });
       }
 
@@ -196,11 +208,11 @@
           .then(function(similarItem) {
             scope.similarItem = similarItem;
 
+            // TODO: move something in from the vault to equip!
+            // TODO: do we need this exotic logic?
             // could this be removed now, along with all refrences to `equipExotic` that are passed in?
-            if (!equipExotic && (similarItem) && (similarItem.tier === 'Exotic')) {
-              return $q.reject('There are no items to equip in the \'' + item.type + '\' slot.');
-            } else if (!similarItem) {
-              return $q.reject('There are no items to equip in the \'' + item.type + '\' slot.');
+            if ((!equipExotic && similarItem && similarItem.tier === 'Exotic') || !similarItem) {
+              return $q.reject(new Error('There are no items to equip in the \'' + item.type + '\' slot.'));
             }
 
             return dimStoreService.getStore(item.owner);
@@ -226,7 +238,7 @@
                     return dimBungieService.transfer(scope.similarItem, vault);
                   })
                   .then(function() {
-                    updateItemModel(scope.similarItem, vault, scope.source, false);
+                    return updateItemModel(scope.similarItem, vault, scope.source, false);
                   });
               }
 
@@ -234,7 +246,7 @@
                   return dimBungieService.transfer(scope.similarItem, scope.source);
                 })
                 .then(function() {
-                  updateItemModel(scope.similarItem, (vault) ? vault : scope.target, scope.source, false);
+                  return updateItemModel(scope.similarItem, (vault) ? vault : scope.target, scope.source, false);
                 });
             }
           })
@@ -249,14 +261,14 @@
           });
       }
 
-      function moveToVault(item) {
+      function moveToVault(item, amount) {
         return dimStoreService.getStore('vault')
           .then(function(target) {
-            return moveToStore(item, target, false);
+            return moveToStore(item, target, false, amount);
           });
       }
 
-      function moveToStore(item, store, equip) {
+      function moveToStore(item, store, equip, amount) {
         var scope = {
           source: null,
           target: store
@@ -266,14 +278,10 @@
           .then(function(source) {
             scope.source = source;
 
-            // if (_.has(item, 'moveAmount') && (item.moveAmount > 0)) {
-            //   item.amount = item.moveAmount;
-            // }
-
-            return dimBungieService.transfer(item, scope.target);
+            return dimBungieService.transfer(item, scope.target, amount);
           })
           .then(function() {
-            return updateItemModel(item, scope.source, scope.target, false);
+            return updateItemModel(item, scope.source, scope.target, false, amount);
           })
           .then(function(item) {
             if ((item.owner !== 'vault') && equip) {
@@ -362,17 +370,7 @@
           .value();
 
         if (item.maxStackSize > 1) {
-          stackAmount = _(store.items)
-            .chain()
-            .where({
-              hash: item.hash
-            })
-            .pluck('amount')
-            .reduce(function(memo, amount) {
-              return memo + amount;
-            }, 0)
-            .value();
-
+          stackAmount = store.amountOfItem(item);
           slotsNeededForTransfer = Math.ceil((stackAmount + item.amount) / item.maxStackSize) - Math.ceil((stackAmount) / item.maxStackSize);
         } else {
           if (item.owner === store.id) {
@@ -416,7 +414,6 @@
           }
         }
 
-        // TODO Need to add support to transfer partial stacks.
         if ((itemsInStore + slotsNeededForTransfer) <= typeQtyCap) {
           if ((item.owner !== store.id) && (store.id !== 'vault') && (item.owner !== 'vault')) {
             // It's a guardian-to-guardian move, so we need to check
@@ -495,7 +492,7 @@
         });
       }
 
-      function moveTo(item, target, equip) {
+      function moveTo(item, target, equip, amount) {
         var data = {
           item: item,
           source: null,
@@ -521,16 +518,23 @@
             data.target = targetStore;
           })
           .then(function(a) {
-            var promise = $q.when();
+            var promise = $q.when(item);
 
             if (!data.isVault.source && !data.isVault.target) { // Guardian to Guardian
               if (data.source.id != data.target.id) { // Different Guardian
                 if (item.equipped) {
-                  promise = promise.then(dequipItem.bind(null, item));
+                  promise = promise.then(function() {
+                    return dequipItem(item);
+                  });
                 }
 
-                promise = promise.then(moveToVault.bind(null, item))
-                  .then(moveToStore.bind(null, item, data.target, equip));
+                promise = promise
+                  .then(function() {
+                    return moveToVault(item, amount);
+                  })
+                  .then(function(item) {
+                    return moveToStore(item, data.target, equip, amount);
+                  });
               }
 
               if (equip) {
@@ -538,15 +542,15 @@
                   if (!item.equipped) {
                     return equipItem(item);
                   } else {
-                    return $q.when(null);
+                    return $q.when(item);
                   }
                 });
               } else if (!equip) {
                 promise = promise.then(function() {
                   if (item.equipped) {
-                    return dequipItem.bind(null, item)();
+                    return dequipItem(item);
                   } else {
-                    return $q.when(null);
+                    return $q.when(item);
                   }
                 });
               }
@@ -555,16 +559,15 @@
               //console.log('vault-to-vault');
             } else if (data.isVault.source || data.isVault.target) { // Guardian to Vault
               if (item.equipped) {
-                promise = promise.then(dequipItem.bind(null, item));
+                promise = promise.then(function() {
+                  return dequipItem(item);
+                });
               }
 
-              promise = promise.then(moveToStore.bind(null, item, data.target, equip));
-            }
-
-            promise = promise
-              .then(function() {
-                item.moveAmount = 0;
+              promise = promise.then(function() {
+                return moveToStore(item, data.target, equip, amount);
               });
+            }
 
             return promise;
           })
@@ -577,47 +580,15 @@
 
       function getItems() {
         var returnValue = [];
-        var stores = dimStoreService.getStores();
-
-        angular.forEach(stores, function(store) {
+        dimStoreService.getStores().forEach(function(store) {
           returnValue = returnValue.concat(store.items);
         });
-
         return returnValue;
       }
 
-      function getItem(id, hash, amount, store) {
-        var items;
-
-        if (store) {
-          items = store.items;
-        } else {
-          items = getItems();
-        }
-
-        var item;
-
-        if (_.isObject(id)) {
-          var primitive = id;
-
-          item = _.find(items, function(item) {
-            return ((item.id === primitive.id) && (item.hash === primitive.hash));
-          });
-        } else {
-          predicate = {};
-
-          if (!_.isEmpty(id)) {
-            predicate.id = id;
-          }
-
-          if (!_.isEmpty(hash)) {
-            predicate.hash = hash;
-          }
-
-          item = _.findWhere(items, predicate);
-        }
-
-        return item;
+      function getItem(params, store) {
+        var items = store ? store.items : getItems();
+        return _.findWhere(items, { id: params.id, hash: params.hash });
       }
 
       return service;