+++ app/scripts/services/dimLoadoutService.factory.js
@@ -19,14 +19,15 @@
       previousLoadouts: {} // by character ID
     };
 
-    function addItemToLoadout(item) {
+    function addItemToLoadout(item, $event) {
       $rootScope.$broadcast('dim-store-item-clicked', {
-        item: item
+        item: item,
+        clickEvent: $event
       });
     }
 
     function processLoadout(data, version) {
-      if (!_.isUndefined(data)) {
+      if (data) {
         if (version === 'v3.0') {
           var ids = data['loadouts-v3.0'];
           _loadouts.splice(0);
@@ -97,7 +98,7 @@
       var deferred = $q.defer();
       var result;
 
-      if (!_.isUndefined(loadouts)) {
+      if (loadouts) {
         result = $q.when(loadouts);
       } else {
         result = getLoadouts();
@@ -199,118 +200,30 @@
         failed: false
       };
 
-      var items = _.chain(loadout.items)
-        .values()
-        .flatten()
-        .value();
+      var items = _.flatten(_.values(loadout.items));
 
-      var _types = _.chain(items)
-        .pluck('type')
-        .uniq()
-        .value();
+      var _types = _.uniq(_.pluck(items, 'type'));
 
-      var _items = _.chain(store.items)
+      // Existing items on the character, minus ones that we want as part of the loadout
+      var existingItems = _.chain(store.items)
         .filter(function(item) {
-          return _.contains(_types, item.type);
+          return _.contains(_types, item.type) &&
+            (!_.any(items, function(i) {
+              return i.id === item.id && i.hash === item.hash;
+            }));
         })
-        .filter(function(item) {
-          return (!_.some(items, function(i) {
-            return ((i.id === item.id) && (i.hash === item.hash));
-          }));
-        })
-        .groupBy(function(item) {
-          return item.type;
-        })
+        .groupBy('type')
         .value();
 
-      return applyLoadoutItems(store, items, loadout, _items, scope);
+      // TODO: parallelize builds between "weapons" and "general" and "armor", etc
+      return applyLoadoutItems(store, items, loadout, existingItems, scope);
     }
 
-    function applyLoadoutItems(store, items, loadout, _items, scope) {
-      if (items.length > 0) {
-        var pseudoItem = items.splice(0, 1)[0];
-        var item = dimItemService.getItem(pseudoItem);
-
-        if (item.type === 'Class') {
-          item = _.findWhere(store.items, {
-            hash: pseudoItem.hash
-          });
-        }
-
-        if (item) {
-          var size = _.chain(store.items)
-            .filter(function(i) {
-              return item.type === i.type;
-            })
-            .size()
-            .value();
-
-          var p = $q.when(item);
-
-          var target;
-
-          if (size === 10) {
-            if (item.owner !== store.id) {
-              var moveItem = _items[item.type].splice(0, 1);
-              p = $q.when(dimStoreService.getStores())
-                .then(function(stores) {
-                  return _.chain(stores)
-                    .sortBy(function(s) {
-                      if (s.id === store.id) {
-                        return 0;
-                      } else if (s.id === 'vault') {
-                        return 2;
-                      } else {
-                        return 1;
-                      }
-                    })
-                    .value();
-                })
-                .then(function(sortedStores) {
-                  return _.find(sortedStores, function(s) {
-                    return _.chain(s.items)
-                      .filter(function(i) {
-                        return item.type === i.type;
-                      })
-                      .size()
-                      .value() < ((s.id === 'vault') ? 72 : 10);
-                  });
-                })
-                .then(function(t) {
-                  if (_.isUndefined(t)) {
-                    throw new Error('Collector eh?  All your characters ' + moveItem[0].type.toLowerCase() + ' slots are full and I can\'t move items of characters, yet... Clear a slot on a character and I can complete the loadout.');
-                  }
-                  target = t;
-
-                  return dimItemService.moveTo(moveItem[0], target);
-                });
-            }
-          }
-
-          var promise = p
-            .then(function() {
-              return dimItemService.moveTo(item, store, pseudoItem.equipped);
-            })
-            .catch(function(a) {
-              scope.failed = true;
-              toaster.pop('error', item.name, a.message);
-            })
-            .finally(function() {
-              return applyLoadoutItems(store, items, loadout, _items, scope);
-            });
-
-          loadingTracker.addPromise(promise);
-          return promise;
-        } else {
-          return $q.when();
-        }
-      } else {
-        var dimStores;
-
-        return $q.when(dimStoreService.getStores())
-          .then(function(stores) {
-            dimStores = dimStoreService.updateStores(stores);
-          })
+    // Move one loadout item at a time. Called recursively to move items!
+    function applyLoadoutItems(store, items, loadout, existingItems, scope) {
+      if (items.length == 0) {
+        // We're done!
+        return dimStoreService.updateStores(dimStoreService.getStores())
           .then(function() {
             var value = 'success';
             var message = 'Your loadout has been transfered.';
@@ -323,6 +236,103 @@
             toaster.pop(value, loadout.name, message);
           });
       }
+
+      var promise = $q.when();
+      var pseudoItem = items.shift();
+      var item = dimItemService.getItem(pseudoItem);
+
+      if (item.type === 'Material' || item.type === 'Consumable') {
+        // handle consumables!
+        var amountNeeded = pseudoItem.amount - store.amountOfItem(pseudoItem);
+        if (amountNeeded > 0) {
+          var otherStores = _.reject(dimStoreService.getStores(), function(otherStore) {
+            return store.id == otherStore.id;
+          });
+          var storesByAmount = _.sortBy(otherStores.map(function(store) {
+            return {
+              store: store,
+              amount: store.amountOfItem(pseudoItem)
+            };
+          }), 'amount').reverse();
+
+          while (amountNeeded > 0) {
+            var source = _.max(storesByAmount, 'amount');
+            var amountToMove = Math.min(source.amount, amountNeeded);
+            var sourceItem = _.findWhere(source.store.items, { hash: pseudoItem.hash });
+
+            if (amountToMove === 0 || !sourceItem) {
+              promise = promise.then(function() {
+                return $q.reject(new Error("There's not enough " + item.name + " to fulfill your loadout."));
+              });
+              break;
+            }
+
+            source.amount -= amountToMove;
+            amountNeeded -= amountToMove;
+
+            promise = promise.then(function() {
+              return dimItemService.moveTo(sourceItem, store, false, amountToMove);
+            });
+          }
+        }
+      } else {
+        if (item.type === 'Class') {
+          item = _.findWhere(store.items, {
+            hash: pseudoItem.hash
+          });
+        }
+
+        if (item) {
+          var size = _.where(store.items, { type: item.type }).length;
+
+          // If full, make room. TODO: put this in the move service!
+          if (size === 10) {
+            if (item.owner !== store.id) {
+              // TODO: should probably choose the least-desirable item
+              var moveAwayItem = existingItems[item.type].shift();
+              var sortedStores = _.sortBy(dimStoreService.getStores(), function(s) {
+                if (s.id === store.id) {
+                  return 0;
+                } else if (s.id === 'vault') {
+                  return 1;
+                } else {
+                  return 2;
+                }
+              });
+              // Get the first store with space
+              // TODO: handle consumable capacity
+              var target = _.find(sortedStores, function(s) {
+                var capacity = (s.id === 'vault' ? 71 : 10);
+                return _.where(s.items, { type: item.type }).length < capacity;
+              });
+              if (!target) {
+                promise = $q.reject(new Error("Collector, eh?  All your characters' " + moveAwayItem.type.toLowerCase() + ' slots are full and I can\'t move items off characters, yet... Clear a slot on a character and I can complete the loadout.'));
+              } else {
+                promise = dimItemService.moveTo(moveAwayItem, target);
+              }
+            }
+          }
+
+          promise = promise.then(function() {
+            return dimItemService.moveTo(item, store, pseudoItem.equipped);
+          });
+        } else {
+          promise = $.reject(new Error(item.name + " doesn't exist in your account."));
+        }
+      }
+
+      promise = promise
+        .catch(function(e) {
+          scope.failed = true;
+          toaster.pop('error', item.name, e.message);
+        })
+        .finally(function() {
+          // Keep going
+          return applyLoadoutItems(store, items, loadout, existingItems, scope);
+        });
+
+      loadingTracker.addPromise(promise);
+      return promise;
     }
 
     function hydratev3d0(loadoutPrimitive) {
@@ -335,7 +345,7 @@
       };
 
       _.each(loadoutPrimitive.items, function(itemPrimitive) {
-        var item = _.clone(dimItemService.getItem({
+        var item = angular.copy(dimItemService.getItem({
           id: itemPrimitive.id,
           hash: itemPrimitive.hash
         }));
@@ -344,6 +354,8 @@
           var discriminator = item.type.toLowerCase();
 
           item.equipped = itemPrimitive.equipped;
+
+          item.amount = itemPrimitive.amount;
 
           result.items[discriminator] = (result.items[discriminator] || []);
           result.items[discriminator].push(item);
@@ -363,7 +375,7 @@
       };
 
       _.each(loadoutPrimitive.items, function(itemPrimitive) {
-        var item = _.clone(dimItemService.getItem({
+        var item = angular.copy(dimItemService.getItem({
           id: itemPrimitive.id,
           hash: itemPrimitive.hash
         }));
@@ -391,7 +403,7 @@
       };
 
       _.each(loadoutPrimitive.items, function(itemPrimitive) {
-        var item = _.clone(dimItemService.getItem(itemPrimitive));
+        var item = angular.copy(dimItemService.getItem(itemPrimitive));
 
         if (item) {
           var discriminator = item.type.toLowerCase();
@@ -416,17 +428,18 @@
         items: []
       };
 
-      _.chain(loadout.items)
+      result.items = _.chain(loadout.items)
         .values()
         .flatten()
-        .each(function(item) {
-          result.items.push({
+        .map(function(item) {
+          return {
             id: item.id,
             hash: item.hash,
             amount: item.amount,
             equipped: item.equipped
-          });
-        });
+          };
+        })
+        .value();
 
       return result;
     }