+++ app/scripts/store/dimStoreItems.directive.js
@@ -47,7 +47,6 @@
       controller: StoreItemsCtrl,
       controllerAs: 'vm',
       bindToController: true,
-      link: Link,
       replace: true,
       scope: {
         'store': '=storeData'
@@ -62,15 +61,15 @@
         '      </div>',
         '      <div ng-repeat="type in ::value track by type" class="sub-section"',
         '           ng-class="[\'sort-\' + type.replace(\' \', \'-\').toLowerCase(), { empty: !vm.data[vm.orderedTypes[type]] }]"',
-        '           ui-on-drop="vm.onDrop($data, $event, false)"',
+        '           ui-on-drop="vm.onDrop($data, $event, false)" ui-on-drag-enter="vm.onDragEnter($event)" ui-on-drag-leave="vm.onDragLeave($event)"',
         '           drop-channel="{{:: type + \',\' + vm.store.id + type }}">',
         '        <div ng-class="vm.styles[type.replace(\' \', \'-\')].equipped"',
         '             ng-if="::vm.store.id !== \'vault\'"',
-        '             ui-on-drop="vm.onDrop($data, $event, true)"',
+        '             ui-on-drop="vm.onDrop($data, $event, true)" ui-on-drag-enter="vm.onDragEnter($event)" ui-on-drag-leave="vm.onDragLeave($event)"',
         '              drop-channel="{{:: type + \',\' + vm.store.id + type }}">',
         '          <div ng-repeat="item in vm.data[vm.orderedTypes[type]] | equipped:true track by item.index" dim-store-item store-data="vm.store" item-data="item"></div>',
         '        </div>',
-        '        <div ng-class="vm.styles[type.replace(\' \', \'-\')].unequipped" ui-on-drop="vm.onDrop($data, $event, false)" drop-channel="{{ type + \',\' + vm.store.id + type }}">',
+        '        <div ng-class="vm.styles[type.replace(\' \', \'-\')].unequipped" ui-on-drop="vm.onDrop($data, $event, false)" ui-on-drag-enter="vm.onDragEnter($event)" ui-on-drag-leave="vm.onDragLeave($event)" drop-channel="{{ type + \',\' + vm.store.id + type }}">',
         '          <div ng-repeat="item in vm.data[vm.orderedTypes[type]] | equipped:false | sortItems:vm.itemSort track by item.index" dim-store-item store-data="vm.store" item-data="item"></div>',
         '          <div class="item-target"></div>',
         '        </div>',
@@ -80,24 +79,48 @@
         '</div>'
       ].join('')
     };
-
-    function Link(scope, element, attrs) {
-      scope.vm.onDrop = function(id, $event, equip) {
-        var vm = scope.vm;
-
-        var srcElement = $('#' + id);
-
-        vm.moveDroppedItem(angular.element(srcElement[0])
-          .scope()
-          .item, equip);
-      };
-    }
   }
 
-  StoreItemsCtrl.$inject = ['$scope', 'loadingTracker', 'dimStoreService', 'dimItemService', '$q', '$timeout', 'toaster', 'dimSettingsService'];
 
-  function StoreItemsCtrl($scope, loadingTracker, dimStoreService, dimItemService, $q, $timeout, toaster, dimSettingsService) {
+  StoreItemsCtrl.$inject = ['$scope', 'loadingTracker', 'dimStoreService', 'dimItemService', '$q', '$timeout', 'toaster', 'dimSettingsService', 'ngDialog', '$rootScope'];
+
+  function StoreItemsCtrl($scope, loadingTracker, dimStoreService, dimItemService, $q, $timeout, toaster, dimSettingsService, ngDialog, $rootScope) {
     var vm = this;
+
+    // Detect when we're hovering a dragged item over a target
+    var dragTimer = null;
+    var hovering = false;
+    var dragHelp = document.getElementById('drag-help');
+    var entered = 0;
+    vm.onDragEnter = function($event) {
+      if ($rootScope.dragItem && $rootScope.dragItem.owner !== vm.store.id) {
+        entered = entered + 1;
+        if (entered === 1) {
+          dragTimer = $timeout(function() {
+            if ($rootScope.dragItem) {
+              hovering = true;
+              dragHelp.classList.add('drag-dwell-activated');
+            }
+          }, 1000);
+        }
+      }
+    };
+    vm.onDragLeave = function($event) {
+      if ($rootScope.dragItem && $rootScope.dragItem.owner !== vm.store.id) {
+        entered = entered - 1;
+        if (entered === 0) {
+          hovering = false;
+          dragHelp.classList.remove('drag-dwell-activated');
+          $timeout.cancel(dragTimer);
+        }
+      }
+    };
+    vm.onDrop = function(id, $event, equip) {
+      vm.moveDroppedItem(angular.element('#' + id).scope().item, equip, $event);
+      hovering = false;
+      dragHelp.classList.remove('drag-dwell-activated');
+      $timeout.cancel(dragTimer);
+    };
 
     var types = [ // Order of types in the rows.
       'Class',
@@ -254,6 +277,18 @@
         equipped: '',
         unequipped: 'unequipped equippable',
       },
+      Bounties: {
+        equipped: '',
+        unequipped: 'unequipped equippable',
+      },
+      Quests: {
+        equipped: '',
+        unequipped: 'unequipped equippable',
+      },
+      Missions: {
+        equipped: '',
+        unequipped: 'unequipped equippable',
+      },
       'Special-Orders': {
         equipped: '',
         unequipped: 'unequipped equippable',
@@ -264,47 +299,105 @@
       }
     };
 
-    // TODO: Consolidate this with the same code in dimMovePopup.directive.js
-    vm.moveDroppedItem = function(item, equip) {
-      var promise = null;
+    vm.moveDroppedItem = function(item, equip, $event) {
       var target = vm.store;
+
+      if (item.notransfer && item.owner !== target.id) {
+        return $q.reject(new Error('Cannot move that item off this character.'));
+      }
 
       if (item.owner === vm.store.id) {
         if ((item.equipped && equip) || (!item.equipped) && (!equip)) {
-          return $q.resolve();
+          return $q.resolve(item);
         }
-
-        promise = $q.when(vm.store);
-
-      } else {
-        promise = dimStoreService.getStore(item.owner);
       }
 
-      if (item.notransfer && item.owner !== target.id) {
-        promise = $q.reject(new Error('Cannot move that item off this character.'));
-      }
+      var promise = $q.when(item.amount);
 
-      var dimStores = null;
-
-      var reload = item.equipped || equip;
-
-      promise = promise.then(dimItemService.moveTo.bind(null, item, target, equip));
-
-      if (reload) {
-        promise = promise.then(dimStoreService.getStores)
-          .then(function(stores) {
-            dimStores = dimStoreService.updateStores(stores);
-          });
-      }
-      promise = promise
-        .then(function() {
-          setTimeout(function() { dimStoreService.setHeights(); }, 0);
-        })
-        .catch(function(a) {
-          toaster.pop('error', item.name, a.message);
+      if (item.maxStackSize > 1 && item.amount > 1 && ($event.shiftKey || hovering)) {
+        var dialogResult = ngDialog.open({
+          // TODO: break this out into a separate service/directive?
+          template: [
+            '<div>',
+            '  <h1>',
+            '    <dim-infuse-item item-data="vm.item"></dim-infuse-item>',
+            '    How much {{vm.item.name}} to move?',
+            '  </h1>',
+            '  <div class="ngdialog-inner-content">',
+            '    <form ng-submit="vm.finish()">',
+            '      <dim-move-amount amount="vm.moveAmount" maximum="vm.maximum"></dim-move-amount>',
+            '    </form>',
+            '    <div class="buttons"><button ng-click="vm.finish()">Move</button></buttons>',
+            '  </div>',
+            '</div>'].join(''),
+          scope: $scope,
+          resolve: {
+            maximum: function() {
+              return dimStoreService.getStore(item.owner)
+                .then(function(store) {
+                  return store.amountOfItem(item);
+                });
+            }
+          },
+          controllerAs: 'vm',
+          controller: ['$scope', 'maximum', function($scope, maximum) {
+            var vm = this;
+            vm.item = $scope.ngDialogData;
+            vm.moveAmount = vm.item.amount;
+            vm.maximum = maximum;
+            vm.finish = function() {
+              $scope.closeThisDialog(vm.moveAmount);
+            };
+          }],
+          plain: true,
+          data: item,
+          appendTo: 'body',
+          overlay: true,
+          className: 'move-amount-popup'
         });
 
+        promise = dialogResult.closePromise.then(function(data) {
+          if (typeof data.value === 'string') {
+            return $q.reject(new Error("move-canceled"));
+          }
+          var moveAmount = data.value;
+          return moveAmount;
+        });
+      }
+
+      promise.then(function(moveAmount) {
+        var getStore;
+        if (item.owner === vm.store.id) {
+          getStore = $q.when(vm.store);
+        } else {
+          getStore = dimStoreService.getStore(item.owner);
+        }
+        var dimStores = null;
+
+        var movePromise = getStore.then(function() {
+          return dimItemService.moveTo(item, target, equip, moveAmount);
+        });
+
+        var reload = item.equipped || equip;
+        if (reload) {
+          movePromise = movePromise
+            .then(dimStoreService.getStores)
+            .then(function(stores) {
+              dimStores = dimStoreService.updateStores(stores);
+            });
+        }
+        return movePromise.then(function() {
+          setTimeout(function() { dimStoreService.setHeights(); }, 0);
+        });
+      }).catch(function(e) {
+        if (e.message !== 'move-canceled') {
+          toaster.pop('error', item.name, e.message);
+        }
+      });
+;
+
       loadingTracker.addPromise(promise);
+
       return promise;
     };